Add files via upload
I have been learning Flowcharts and Pseudocodes this week in ALX, and I managed to use the if else statements in EXCEL making the work easier, the pass mark used was 50%. The question was to find the students who will get admitted to Analytical Chemistry pHD programme given the conditions that they must pass both Computational Chemistry and Medicinal Chemistry and either one of Organic Chemistry and S&P Chemistry.  That's my EXCEL sheet and Pseudocode that I came up with.